+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0
 - Add `cigar.to_char_pairs(pattern, text)` to iterate corresponding characters.
+- Add `cigar.clear()` for `cigar.ops.clear()`.
 
 ## 1.1.0
 - #2: `cigar.to_string()` explicitly writes counts of 1.